+++ src/pages/imageSetup/apis/activityInfo.ts
