+++ src/pages/imageSetup/apis/floorPlan.ts
