+++ src/pages/imageSetup/apis/houseInfo.ts
