+++ src/pages/imageSetup/apis/interiorStyle.ts
